fix priority

+++ supla-console-client/src/client_config.cpp
@@ -99,7 +99,7 @@ bool client_config::load(const char* config_file) {
             command["message"].As<std::string>("sample message"),
             pushover_token, pushover_user, reset,
             command["command"].As<std::string>(""),
-            command["priority"].As<uint8_t>(0));
+            command["priority"].As<int>(0));
       }
     }
